requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
celery==5.5.0
redis==6.1.0
SQLAlchemy==2.0.43
python-dotenv==1.1.0
pydantic==2.4.0
requests==2.32.0

